Question 1

+++ C++/project 1.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 int main(){
-	int temp,frani;
+	double temp,frani;
 	cout<<"Enter the Temp :"<<endl;
 	cin>>temp;
 	frani=(temp*1.8)+32;
